Updated seed.

+++ db/seeds.rb
@@ -2,12 +2,14 @@ ActiveRecord::Base.establish_connection
 ActiveRecord::Base.connection.tables.each do |table|
   next if table == 'schema_migrations'
 
-  # MySQL and PostgreSQL
-  # ActiveRecord::Base.connection.execute("TRUNCATE #{table}")
-
-  # SQLite
-  ActiveRecord::Base.connection.execute("DELETE FROM #{table}")
-  ActiveRecord::Base.connection.execute("DELETE FROM SQLITE_SEQUENCE WHERE NAME = '#{table}'")
+  if if Rails.env.production?
+    # MySQL and PostgreSQL
+    ActiveRecord::Base.connection.execute("TRUNCATE #{table}")
+  else
+    # SQLite
+    ActiveRecord::Base.connection.execute("DELETE FROM #{table}")
+    ActiveRecord::Base.connection.execute("DELETE FROM SQLITE_SEQUENCE WHERE NAME = '#{table}'")
+  end
 end
 
 3.times do
